pyproject.toml
[project]
name = "restaurant-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.11.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
]
